+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # JEC GlassCat Metadata Project: Update Release Notes
 
+<a name="jec-glasscat-metadata-1.0.4"></a>
+## **1.0.4** (2018-08-30)
+
+### Bug Fixes
+
+- Fixing metadata validation error when the `creation` property is `null`
+- Fixing test issues due to the `creation` property changes
+
+### Features
+
+- Updating all dependencies
+- Test cases improvements
+
+<a name="jec-glasscat-metadata-1.0.3"></a>
+## **1.0.3** (2018-08-22)
+
+### Bug Fixes
+
+### Features
+
+- Dependencies upgrade
+
 <a name="jec-glasscat-metadata-1.0.2"></a>
 ## **1.0.2** (2018-03-28)